Created routes and added body-parser
Created  /add-product route which shows a form. Once user has filled , consoled it after parsing it using body-parser. Also added one more input tag in the add product page so that user can input size of the product. When the user clicks on submit parsed both the input tags and showed it in console.